+++ apps/docs/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @chanchan2/docs
+# @blackstarzck/docs
 
 ## 0.1.1
 
@@ -6,5 +6,5 @@
 
 - Updated dependencies [70b45b7]
 - Updated dependencies [70b45b7]
-  - @chanchan2/ui@0.2.0
-  - @chanchan2/tokens@0.2.0
+  - @blackstarzck/ui@0.2.0
+  - @blackstarzck/tokens@0.2.0